// apps/informative-site/sanity.types.ts
export type LinkInternal = {
  _type: "linkInternal";
  label?: string;
  reference?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "page";
  } | {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "home";
  };
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SettingsQueryResult = {
  _id: string;
  _type: "settings";
  footer: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }> | null;
  menuItems: Array<{
    _key: null;
    _type: "home";
    slug: null;
    title: string | null;
  } | {
    _key: null;
    _type: "page";
    slug: string | null;
    title: string | null;
  } | {
    _key: null;
    _type: "project";
    slug: string | null;
    title: string | null;
  }> | null;
  nav: {
    menus: Array<{
      title?: string;
      submenus: Array<{
        link?: LinkInternal;
        _type: "submenuItem";
        _key: string;
        slug: string | null;
        type_reference: "home" | "page" | null;
      }> | null;
      _type: "menuWithSubmenu";
      _key: string;
      slug: null;
      type_reference: null;
    } | {
      link?: LinkInternal;
      _type: "simpleMenu";
      _key: string;
      slug: string | null;
      type_reference: "home" | "page" | null;
      submenus: null;
    }> | null;
  } | null;
  ogImage: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
} | null;